requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
curl_cffi==0.11.3
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google-search-results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.16.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.42.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
openai-whisper==20240930
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytube==15.0.0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
schedule==1.2.2
scipy==1.15.3
selenium==4.33.0
serpapi==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SoundCard==0.4.4
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1+cpu
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.62
